Use openapi generate from node
The project was originally dependent on openapi being installed, but
that was causing issues finding the relevant version to install.

To make it slightly easier to set up the environment, this commit
uses `@openapitools/openapi-generator-cli` to fetch the relevant
version. We're stopping at 6.6.0 as later versions have issues with
the project. We also need Java 15 or earlier, as Java 16 closes access
to a number of classes that OpenAPI uses (via the Java Moustache
library).